+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/security/ApiKeyAuth.kt
@@ -2,6 +2,8 @@ package io.bkbn.kompendium.oas.security
 
 import java.util.Locale
 
+// TODO... is there even an official ktor api auth mechanism??
+
 @Suppress("UnusedPrivateMember")
 class ApiKeyAuth(val `in`: ApiKeyLocation, name: String) : SecuritySchema {
   val type: String = "apiKey"